Setup project on another computer
Setting up the project on another computer. Creating .env files on the front and backend. Setting up the mongoDb. When seeding the User to the db, the user will get a specific ObjectId. That id is stored on the .env file in the backend folder and is then used in the index.js file to get the user by id. Then seed User, on any device i work on, will have the same id so that i dont have to change the id when quering the user by id.