TimeSchedule-fixed bugs
Fixed the problem with selecting of dates. Problem was when endDay key was "null" in the eachOfDayInterval function. That apeared when only one day was selected. Adding a turnery operation to check if the value was null or not. If it was the endDate got the same value as the startDay key. Also added a text that will be rendered if the selected day is a weekend.
In the dayTimePicker, a function will be triggered when the state pushToBeExamined changes to true. The function will call checoutDates() that will clear the selected and timeValue arrays and push selected-items to the toBeExamined array.
When both dates and times is selected, an Alert-success text will appear.